AqXCo 0.53
Modificado:
-camaraComponent
-componente
-playerComponent
-renderComponent
-stateMachineComponent
-transformComponent
-velocityComponent
-facadeMotorGradico
-diccionarioEnCo
-entityManager
-gameClock
-vectorEntity
-camaraSystem
-playerSystem
-main
Incluido:
-handleMoverSystem
-handleMoverComponent

+++ DESARROLLO/HITO1/ITERACION 3/ArquitecturaXComponentes/entitySystem/framework/gameClock.cpp
@@ -76,7 +76,7 @@ unsigned long gameClock::getTime() {
 
 double gameClock::timeElapsed() {
     now = this->getTime();
-    float dt = (now - then) / 1000.f;
+    double dt = (now - then) / 1000.;
     then = now;
     return dt;
 }
